Move modules to separate directory and refactor
