+++ src/Project.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from "react";
-import {ProjectInterface} from "./d";
-import dateConverter from "./DateConverter";
+import {ProjectInterface} from "./Modules/d";
+import dateConverter from "./Modules/DateConverter";
 import ProjectToDoContainer from "./ProjectToDoContainer";
-import DateConverter from "./DateConverter";
+import DateConverter from "./Modules/DateConverter";
 
 export default function Project(props: { project: ProjectInterface, openToDoForm: (project: ProjectInterface) => void,
     deleteProject: (project: ProjectInterface) => void,
